[Feat] MyGroupList 구현함.

+++ travelmaker_android/app/src/main/java/com/gumibom/travelmaker/data/dto/mygroup/MyMeetingGroupDTOItem.kt
@@ -5,7 +5,7 @@ data class MyMeetingGroupDTOItem(
     val dday: Int,
     val headId: Int,
     val headNickname: String,
-    val headTrust: Int,
+    val headTrust: Double,
     val isFinish: Boolean,
     val isMeetingFinish: Boolean,
     val mainImgUrl: String,
